+++ R/req-retries.R
@@ -4,17 +4,17 @@
 #' `req_retry()` alters [req_perform()] so that it will automatically retry
 #' in the case of failure. To activate it, you must specify either the total
 #' number of requests to make with `max_tries` or the total amount of time
-#' to spend with `max_seconds`. Then `req_perform()` will retry if:
+#' to spend with `max_seconds`. Then `req_perform()` will retry if the error is
+#' "transient", i.e. it's an HTTP error that can be resolved by waiting. By
+#' default, 429 and 503 statuses are treated as transient, but if the API you
+#' are wrapping has other transient status codes (or conveys transient-ness
+#' with some other property of the response), you can override the default
+#' with `is_transient`.
 #'
-#' * Either the HTTP request or HTTP response doesn't complete successfully
-#'   leading to an error from curl, the lower-level library that httr2 uses to
-#'   perform HTTP request. This occurs, for example, if your wifi is down.
-#'
-#' * The error is "transient", i.e. it's an HTTP error that can be resolved
-#'   by waiting. By default, 429 and 503 statuses are treated as transient,
-#'   but if the API you are wrapping has other transient status codes (or
-#'   conveys transient-ness with some other property of the response), you can
-#'   override the default with `is_transient`.
+#' Additionally, if you set `retry_on_error = TRUE`, the request will retry
+#' if either the HTTP request or HTTP response doesn't complete successfully
+#' leading to an error from curl, the lower-level library that httr2 uses to
+#' perform HTTP request. This occurs, for example, if your wifi is down.
 #'
 #' It's a bad idea to immediately retry a request, so `req_perform()` will
 #' wait a little before trying again:
@@ -42,6 +42,8 @@
 #' @param is_transient A predicate function that takes a single argument
 #'   (the response) and returns `TRUE` or `FALSE` specifying whether or not
 #'   the response represents a transient error.
+#' @param retry_on_error Treat low-level HTTP errors as if they are transient
+#'   errors, and can be retried.
 #' @param backoff A function that takes a single argument (the number of failed
 #'   attempts so far) and returns the number of seconds to wait.
 #' @param after A function that takes a single argument (the response) and
@@ -80,16 +82,19 @@
 req_retry <- function(req,
                       max_tries = NULL,
                       max_seconds = NULL,
+                      retry_on_error = FALSE,
                       is_transient = NULL,
                       backoff = NULL,
                       after = NULL) {
   check_request(req)
   check_number_whole(max_tries, min = 2, allow_null = TRUE)
   check_number_whole(max_seconds, min = 0, allow_null = TRUE)
+  check_bool(retry_on_error)
 
   req_policies(req,
     retry_max_tries = max_tries,
     retry_max_wait = max_seconds,
+    retry_on_error = retry_on_error,
     retry_is_transient = as_callback(is_transient, 1, "is_transient"),
     retry_backoff = as_callback(backoff, 1, "backoff"),
     retry_after = as_callback(after, 1, "after")
@@ -106,6 +111,10 @@ retry_max_seconds <- function(req) {
 }
 
 retry_is_transient <- function(req, resp) {
+  if (is_error(resp)) {
+    return(req$policies$retry_on_error %||% FALSE)
+  }
+
   req_policy_call(req, "retry_is_transient", list(resp),
     default = function(resp) resp_status(resp) %in% c(429, 503)
   )
@@ -116,6 +125,10 @@ retry_backoff <- function(req, i) {
 }
 
 retry_after <- function(req, resp, i, error_call = caller_env()) {
+  if (is_error(resp)) {
+    return(retry_backoff(req, i))
+  }
+
   after <- req_policy_call(req, "retry_after", list(resp), default = resp_retry_after)
 
   # TODO: apply this idea to all callbacks